requirements-dev.txt
-rrequirements.txt
mock
pytest
pytest-coverage
vcrpy
